pull-split-helper.py (#45)
* inital move of clearly assignable  functions

* remove not needed functions

* fixed false comment traceback

* moved now-sure functions

* moved import sections

* reduced imports to relevant and fixed format

* linfit helpers tests - adjusted import

* reformat

* cut all tests

* put all relevant tests here

* docstring top

* remove import not relevant

* almost forgot about testing mpc

* renamed (_ prefix) to internal/not to use module - done
